add unit tests

+++ src/etc/utils.ts
@@ -6,9 +6,9 @@
 
 import path from 'path';
 import fs from 'fs-extra';
-import * as pack from '../package.json';
-import { semanticGroups, versionsOptions } from './types';
-import { Logger, TypeDocReader } from 'typedoc';
+import * as pack from '../../package.json';
+import { minorVersion, patchVersion, semanticAlias, semanticGroups, version, versionsOptions } from '../types';
+import { Application, Logger, TypeDocReader } from 'typedoc';
 
 /**
  * Attempts to find the github repository location url and parse it into a github pages url. 
@@ -19,7 +19,7 @@ export function getGhPageUrl(repository: {url} = pack.repository) {
 	if (!repository || !repository.url) {
 		return 'http://localhost:5500/docs';
 	}
-	const splitUrl = pack.repository.url.split('/'); 
+	const splitUrl = repository.url.split('/'); 
 	const gitName = splitUrl[3];
 	const repoName = splitUrl[4];
 	return `https://${gitName}.github.io/${repoName}`;
@@ -30,7 +30,7 @@ export function getGhPageUrl(repository: {url} = pack.repository) {
  * @param version
  * @returns The package version
  */
-export function getPackageVersion(version: string = pack.version): string{
+export function getPackageVersion(version: string | null = pack.version): string{
 	!version && (()=>{throw new Error('Package version was not found')})();
 	return version;
 }
@@ -46,6 +46,18 @@ export function getMinorPackageVersion(){
 }
 
 /**
+ * Parses the root document directory for all semantically named sub-directories.
+ * @param docRoot 
+ * @returns an array of directories
+ */
+ export function getPackageDirectories(docRoot): [string]{
+	return fs.readdirSync(docRoot).filter(file => {		
+		const stats = fs.statSync(path.join(docRoot, file));
+		return stats.isDirectory() && /^v[\d]+.[\d]+.[\d]+/.test(file)
+	}) as [string];
+}
+
+/**
  * Creates groups of semantic versions with the latest patch version identified
  * @param directories an array of semantically named directories to be processed
  * @returns 
@@ -56,7 +68,6 @@ export function getSemGroups(directories: [string]): semanticGroups {
 		let [major, minor, patch] = dir.split('.') as [string, string, number];
 		if (typeof patch !== 'undefined') {
 			patch = patch*1;
-			console.log(major, minor, patch);
 			!semGroups[major] && (semGroups[major] = {});
 			(!semGroups[major][minor] || semGroups[major][minor] < patch) &&
 				(semGroups[major][minor] = patch)
@@ -95,53 +106,58 @@ export const DOC_VERSIONS = [
  * @param docRoot The root url of the documentation site
  * @returns a string of valid HTML 
  */
-export function makeIndex(docRoot){
+export function makeIndex(docRoot: string){
 	const baseUrl = new URL(docRoot);
 	const newUrl = new URL(path.join(baseUrl.pathname, 'stable'), baseUrl.origin)
 	return `<meta http-equiv="refresh" content="0; url=${newUrl.href}"/>`
 }
 
-/**
- * Parses the root document directory for all semantically named sub-directories.
- * @param oldOut 
- * @returns an array of directories
- */
-export function getPackageDirectories(oldOut): [string]{
-	return fs.readdirSync(oldOut).filter(file => {
-		const stats = fs.statSync(path.join(oldOut, file));
-		return stats.isDirectory() && /^v[\d]+.[\d]+.[\d]+/.test(file)
-	}) as [string];
-}
 
 /**
- * Creates a symlink for a non-semantically named alias release 
- * @param rootPath 
+ * Creates a symlink for the stable release 
+ * @param docRoot 
  * @param semGroups 
- * @param version 
+ * @param pegVersion 
  * @param name 
  */
-export function makePeggedLink(rootPath: string, semGroups:semanticGroups, version: string, name: string): void {
-	const stableArray = version.split('.');
+export function makeStableLink(docRoot: string, semGroups:semanticGroups, pegVersion: minorVersion, name: semanticAlias = 'stable'): void {
+	const stableArray = pegVersion.split('.');
 	let stableSource = `v${stableArray[0]}.${stableArray[1]}.${semGroups['v'+stableArray[0]][stableArray[1]]}`;
-	stableSource = path.join(rootPath, stableSource);
-	const stableTarget = path.join(rootPath, name);
+	stableSource = path.join(docRoot, stableSource);
+	if (!fs.existsSync(stableSource)) throw new Error(`Document directory does not exist: ${stableSource}`);
+	const stableTarget = path.join(docRoot, name);
 	fs.existsSync(stableTarget) && fs.unlinkSync(stableTarget);
 	fs.createSymlinkSync(stableSource, stableTarget, 'dir')
 }
 
 /**
+ * Creates a symlink for the dev release 
+ * @param docRoot 
+ * @param semGroups 
+ * @param pegVersion 
+ * @param name 
+ */
+ export function makeDevLink(docRoot: string, semGroups:semanticGroups, pegVersion: patchVersion, name: semanticAlias = 'dev'): void {
+	let devSource = `v${pegVersion}`;
+	devSource = path.join(docRoot, devSource);
+	if (!fs.existsSync(devSource)) throw new Error(`Document directory does not exist: ${devSource}`);
+	const devTarget = path.join(docRoot, name);
+	fs.existsSync(devTarget) && fs.unlinkSync(devTarget);
+	fs.createSymlinkSync(devSource, devTarget, 'dir')
+}
+
+
+/**
  * Creates symlinks for minor versions pointing to the latest patch release
  * @param semGroups 
- * @param rootPath 
+ * @param docRoot 
  */
-export function makeMinorVersionLinks(semGroups, rootPath): void {
+export function makeMinorVersionLinks(semGroups, docRoot): void {
 	Object.keys(semGroups).forEach(major => {
 		Object.keys(semGroups[major]).forEach(minor => {
 			const patch = semGroups[major][minor];
-			const target = path.join(rootPath, `${major}.${minor}`);
-			const src = path.join(rootPath, `${major}.${minor}.${patch}`);
-			console.log(src, target);
-
+			const target = path.join(docRoot, `${major}.${minor}`);
+			const src = path.join(docRoot, `${major}.${minor}.${patch}`);
 			fs.existsSync(target) && fs.unlinkSync(target);
 			fs.createSymlinkSync(src, target, 'dir');
 		})
@@ -154,10 +170,29 @@ export function makeMinorVersionLinks(semGroups, rootPath): void {
  * @param app 
  * @returns correctly overridden options
  */
-export function bugWorkaround(app): versionsOptions{
+export function getVersionsOptions(app: Application): versionsOptions{
 	const defaultOpts = app.options.getValue('versions') as versionsOptions;
 	app.options.addReader(new TypeDocReader());
 	app.options.read(new Logger());
 	const options = app.options.getValue('versions') as versionsOptions;
 	return {...defaultOpts, ...options };
+}
+
+export function getPaths(app, version?: patchVersion){
+	const rootPath = app.options.getValue('out'); 
+	return {
+		rootPath,
+		targetPath: path.join(rootPath, `v${getPackageVersion(version)}`)
+	}
+}
+export function handleJeckyll(rootPath: string, targetPath: string): void{
+	const srcJeckPath = path.join(targetPath, '.nojekyll');
+	const targetJeckPath = path.join(rootPath, '.nojekyll');
+	fs.existsSync(targetJeckPath) && fs.removeSync(targetJeckPath);
+	fs.existsSync(srcJeckPath) && fs.moveSync(srcJeckPath, targetJeckPath);
+}
+export function handleAssets(targetPath: string){
+	const sourceAsset = path.join(process.cwd(), 'src/assets/versionsMenu.js');
+	fs.ensureDirSync(path.join(targetPath, 'assets'));
+	fs.copyFileSync(sourceAsset, path.join(targetPath, 'assets/versionsMenu.js'));
 }
